Work on "Set video offset" Related to #111

+++ src/UI/Logic/Config/SeGeneral.cs
@@ -93,7 +93,7 @@ public class SeGeneral
     public bool SelectCurrentSubtitleWhilePlaying { get; set; }
     public bool WriteAn2Tag { get; set; }
 
-    public double CurrentVideoOffsetInMs = 0;
+    public long CurrentVideoOffsetInMs = 0;
     public bool CurrentVideoIsSmpte = false;
 
     public SeGeneral()
